removing useless comments, renaming getServerInfoHTML parameter, moved anguar-inview.js and added licence header
